REF: Move service registry code into `octue.cloud.registry` module
